gvr-360video: pause and resume improvements, cleanup
